Correzione bug

+++ JAVA/NewDMApp/src/main/java/db/ScontrinoDao.java
@@ -47,10 +47,10 @@ public class ScontrinoDao {
     }
   }
 
-    public static void checkScontrino(long codice, String dataScontrino) throws ScontrinoException, DatabaseException {
+    public static void checkScontrino(int codice, String dataScontrino) throws ScontrinoException, DatabaseException {
       try {
         PreparedStatement prep = DatabaseConnection.con.prepareStatement(Query.checkScontrino);
-        prep.setLong(1, codice);
+        prep.setInt(1, codice);
         prep.setString(2, dataScontrino);
 
         ResultSet res = prep.executeQuery();
